+++ packages/react/src/exampleTypes/TagsQuery.js
@@ -1,21 +1,108 @@
 import React from 'react'
 import _ from 'lodash/fp'
+import F from 'futil-js'
 import { observer } from 'mobx-react'
 import injectTreeNode from '../utils/injectTreeNode'
 import { exampleTypes } from 'contexture-client'
 import { bgJoin } from '../styles/generic'
 import { TagsInput as DefaultTagsInput } from '../layout/TagsInput'
+import DefaultRadioList from '../layout/RadioList'
+import DefaultSelect from '../layout/Select'
+import TagsJoinPicker, { tagToGroupJoin } from './TagsJoinPicker'
+let CheckboxDefault = props => <input type="checkbox" {...props} />
 
-let tagToGroupJoin = x =>
-  ({
-    any: 'or',
-    all: 'and',
-    none: 'not',
-  }[x])
 let tagValueField = 'word'
-let TagsQuery = injectTreeNode(
-  observer(({ tree, node, TagsInput = DefaultTagsInput, placeholder }) => (
+let TagsQuery = ({
+  tree,
+  node,
+  TagsInput = DefaultTagsInput,
+  Checkbox = CheckboxDefault,
+  RadioList = DefaultRadioList,
+  Select = DefaultSelect,
+  Button = 'button',
+  placeholder,
+}) => {
+  let getTag = tag => _.find({ [tagValueField]: tag }, node.tags)
+  let TagQueryPopever = observer(({ tag }) => {
+    let tagInstance = getTag(tag)
+    return (
+      <div className="tags-input-popover">
+        <div>
+          <div className="popover-item">
+            Keyword: <span className="filter-field-label">{tag}</span>
+          </div>
+          {_.includes(' ', tag) && (
+            <React.Fragment>
+              <div className="popover-item">
+                <RadioList
+                  options={F.autoLabelOptions(['fuzzy', 'exact'])}
+                  value={tagInstance.distance ? 'fuzzy' : 'exact'}
+                  onChange={value => {
+                    tagInstance.distance = value === 'fuzzy' ? 3 : 0
+                    tree.mutate(node.path, { tags: [...node.tags] })
+                  }}
+                />
+              </div>
+              <div className="popover-item">
+                <Button
+                  onClick={() => {
+                    tree.mutate(node.path, {
+                      tags: _.map(tag => {
+                        if (_.includes(' ', tag[tagValueField]))
+                          tag.distance = tagInstance.distance
+                        return tag
+                      }, node.tags),
+                    })
+                  }}
+                >
+                  Apply to all keywords
+                </Button>
+              </div>
+            </React.Fragment>
+          )}
+          <label className="popover-item labeled-checkbox">
+            <Checkbox
+              checked={tagInstance.onlyShowTheseResults}
+              onChange={e => {
+                tagInstance.onlyShowTheseResults = e.target.checked
+                tree.mutate(node.path, { tags: [...node.tags] })
+              }}
+            />
+            <span>Only view this keyword</span>
+          </label>
+        </div>
+        <div>
+          <label className="popover-item labeled-checkbox">
+            <Checkbox
+              checked={!node.exact}
+              onChange={e =>
+                tree.mutate(node.path, { exact: !e.target.checked })
+              }
+            />
+            <span>Enable stemming</span>
+          </label>
+          <div className="popover-item">
+            <TagsJoinPicker node={node} tree={tree} Select={Select} />
+          </div>
+        </div>
+      </div>
+    )
+  })
+  let tagStyle = tag => {
+    let tagInstance = getTag(tag)
+    return {
+      ...(tagInstance.distance ? {} : { fontWeight: 'bold' }),
+      ...bgJoin(tagToGroupJoin(node.join)),
+      opacity:
+        tagInstance.onlyShowTheseResults ||
+        !_.find('onlyShowTheseResults', node.tags)
+          ? 1
+          : 0.5,
+    }
+  }
+  return (
     <TagsInput
+      splitCommas
       tags={_.map(tagValueField, node.tags)}
       addTag={tag => {
         tree.mutate(node.path, {
@@ -27,13 +114,12 @@ let TagsQuery = injectTreeNode(
           tags: _.reject({ [tagValueField]: tag }, node.tags),
         })
       }}
-      tagStyle={bgJoin(tagToGroupJoin(node.join || 'any'))}
+      tagStyle={tagStyle}
       submit={tree.triggerUpdate}
       placeholder={placeholder}
+      PopoverContents={TagQueryPopever}
     />
-  )),
-  exampleTypes.tagsQuery
-)
-TagsQuery.displayName = 'TagsQuery'
+  )
+}
 
-export default TagsQuery
+export default injectTreeNode(observer(TagsQuery), exampleTypes.tagsQuery)