+++ packages/react/src/exampleTypes/index.js
@@ -5,6 +5,7 @@ import DateRangePicker from './DateRangePicker'
 import Query from './Query'
 import Geo from './Geo'
 import TagsQuery from './TagsQuery'
+import TagsText from './TagsText'
 import ResultCount from './ResultCount'
 import ResultTable from './ResultTable'
 import CheckableResultTable from './CheckableResultTable'
@@ -16,6 +17,7 @@ import CheckableTermsStatsTable from './CheckableTermsStatsTable'
 import Text from './Text'
 import { defaultProps } from 'recompose'
 import ModalDefault from '../layout/Modal'
+import DefaultSelect from '../layout/Select'
 
 export default ({
   Input = 'input',
@@ -31,6 +33,7 @@ export default ({
   FieldPicker,
   ListGroupItem = 'div',
   Icon,
+  Select = DefaultSelect,
 } = {}) => {
   let Components = {
     Facet: defaultProps({ TextInput, Checkbox, RadioList })(Facet),
@@ -38,7 +41,9 @@ export default ({
     Date: defaultProps({ DateInput })(Date),
     DateRangePicker,
     Query: defaultProps({ TextInput })(Query),
-    TagsQuery: defaultProps({ TagsInput })(TagsQuery),
+    TagsQuery: defaultProps({ TagsInput, Checkbox, RadioList, Button })(
+      TagsQuery
+    ),
     ResultTable: defaultProps({
       Table,
       Modal,
@@ -54,6 +59,7 @@ export default ({
     CheckableTermsStatsTable,
     Geo,
     Text: defaultProps({ Input })(Text),
+    TagsText: defaultProps({ TagsInput, Select })(TagsText),
   }
   Components.CheckableResultTable = defaultProps({
     ResultTable: Components.ResultTable,
@@ -65,6 +71,7 @@ export default ({
     number: Components.Number,
     date: Components.Date,
     tagsQuery: Components.TagsQuery,
+    tagsText: Components.TagsText,
     geo: Components.Geo,
     text: Components.Text,
   }