+++ packages/react/src/layout/TagsInput.js
@@ -1,11 +1,17 @@
 import React from 'react'
 import _ from 'lodash/fp'
+import F from 'futil-js'
 import { observable } from 'mobx'
 import { observer, inject } from 'mobx-react'
 import { Flex } from './Flex'
+import Popover from './Popover'
 
-let Tag = ({ value, removeTag, tagStyle }) => (
-  <div className="tags-input-tag" style={tagStyle}>
+let Tag = observer(({ value, removeTag, tagStyle, onClick }) => (
+  <div
+    className="tags-input-tag"
+    style={F.callOrReturn(tagStyle, value)}
+    onClick={onClick}
+  >
     {value}
     <span
       className="tags-input-tag-remove"
@@ -18,12 +24,14 @@ let Tag = ({ value, removeTag, tagStyle }) => (
       x
     </span>
   </div>
-)
+))
 Tag.displayName = 'Tag'
 
 let TagsInput = inject(() => ({
   state: observable({
     currentInput: '',
+    selectedTag: null,
+    popoverOpen: false,
   }),
 }))(
   observer(
@@ -36,62 +44,86 @@ let TagsInput = inject(() => ({
       tagStyle,
       TagComponent = Tag,
       placeholder = 'Search...',
-    }) => (
-      <div>
-        <label style={{ display: 'block' }} className="tags-input">
-          <Flex
-            style={{
-              cursor: 'text',
-              alignItems: 'center',
-              flexWrap: 'wrap',
-            }}
-          >
-            {_.map(
-              t => (
-                <TagComponent key={t} value={t} {...{ removeTag, tagStyle }} />
-              ),
-              tags
-            )}
-            <input
-              style={{ border: 'none', outline: 'none', width: 'auto' }}
-              onChange={e => {
-                state.currentInput = e.target.value
+      splitCommas,
+      PopoverContents,
+    }) => {
+      if (splitCommas)
+        addTag = _.flow(
+          _.split(','),
+          _.map(addTag)
+        )
+      return (
+        <div>
+          <label style={{ display: 'block' }} className="tags-input">
+            <Flex
+              style={{
+                cursor: 'text',
+                alignItems: 'center',
+                flexWrap: 'wrap',
               }}
-              onBlur={() => {
-                if (
-                  state.currentInput &&
-                  !_.includes(state.currentInput, tags)
-                ) {
-                  addTag(state.currentInput)
-                  state.currentInput = ''
-                }
-              }}
-              onKeyDown={e => {
-                if (e.key === 'Enter' && !state.currentInput) submit()
-                if (
-                  (e.key === 'Enter' || e.key === 'Tab') &&
-                  state.currentInput &&
-                  !_.includes(state.currentInput, tags)
-                ) {
-                  addTag(state.currentInput)
-                  state.currentInput = ''
-                  e.preventDefault()
-                }
-                if (
-                  e.key === 'Backspace' &&
-                  !state.currentInput &&
-                  tags.length
-                ) {
-                  removeTag(_.last(tags))
-                }
-              }}
-              value={state.currentInput}
-              placeholder={placeholder}
-            />
-          </Flex>
-        </label>
-      </div>
-    )
+            >
+              {_.map(
+                t => (
+                  <TagComponent
+                    key={t}
+                    value={t}
+                    {...{ removeTag, tagStyle }}
+                    onClick={() => {
+                      state.popoverOpen = true
+                      state.selectedTag = t
+                    }}
+                  />
+                ),
+                tags
+              )}
+              <input
+                style={{ border: 'none', outline: 'none', width: 'auto' }}
+                onChange={e => {
+                  state.currentInput = e.target.value
+                }}
+                onBlur={() => {
+                  if (
+                    state.currentInput &&
+                    !_.includes(state.currentInput, tags)
+                  ) {
+                    addTag(state.currentInput)
+                    state.currentInput = ''
+                  }
+                }}
+                onKeyDown={e => {
+                  if (e.key === 'Enter' && !state.currentInput) submit()
+                  if (
+                    (e.key === 'Enter' ||
+                      e.key === 'Tab' ||
+                      (splitCommas && e.key === ',')) &&
+                    state.currentInput &&
+                    !_.includes(state.currentInput, tags)
+                  ) {
+                    addTag(state.currentInput)
+                    state.currentInput = ''
+                    e.preventDefault()
+                  }
+                  if (
+                    e.key === 'Backspace' &&
+                    !state.currentInput &&
+                    tags.length
+                  ) {
+                    removeTag(_.last(tags))
+                  }
+                }}
+                value={state.currentInput}
+                placeholder={placeholder}
+              />
+            </Flex>
+          </label>
+          {PopoverContents && (
+            <Popover isOpen={F.lensProp('popoverOpen', state)}>
+              <PopoverContents tag={state.selectedTag} />
+            </Popover>
+          )}
+        </div>
+      )
+    }
   )
 )
 TagsInput.displayName = 'TagsInput'